Add invoke command & refactor #18
