+++ lamvery/actions/__init__.py
@@ -1,0 +1,9 @@
+from lamvery.actions.archive import ArchiveAction
+from lamvery.actions.configure import ConfigureAction
+from lamvery.actions.decrypt import DecryptAction
+from lamvery.actions.deploy import DeployAction
+from lamvery.actions.encrypt import EncryptAction
+from lamvery.actions.events import EventsAction
+from lamvery.actions.init import InitAction
+from lamvery.actions.set_alias import SetAliasAction
+from lamvery.actions.invoke import InvokeAction
